- Protect from killing stopped processes
git-svn-id: svn://svn.xp-framework.net/xp/trunk@4034 d2cacbed-c0f6-0310-851a-9dad52fc623a

+++ skeleton/lang/Thread.class.php
@@ -164,8 +164,12 @@
      *
      * @access  public
      * @param   int signal default SIGINT
+     * @throws  lang.IllegalThreadStateException
      */
     function stop($signal= SIGINT) {
+      if ($this->_id <= 0) {
+        return throw(new IllegalThreadStateException('Illegal thread id '.$this->_id));
+      }
       posix_kill($this->_id, $signal);
       $this->running= FALSE;
       $this->_id= $this->_pid= -1;
